moved login file to functions folder for organized endpoints
